requirements.txt
# LonaWEB AI - Requirements
# Core dependencies for document analysis and AI chat platform

# Web Framework & UI
streamlit>=1.28.0
streamlit-components>=1.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI & ML Core
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# LLM Integration
llama-cpp-python>=0.2.20

# Vector Database
qdrant-client>=1.6.0

# Document Processing
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing
nltk>=3.8.1
spacy>=3.7.0

# Utilities
pathlib2>=2.3.7
python-multipart>=0.0.6
aiofiles>=23.0.0

# HTTP & Networking
httpx>=0.25.0
requests>=2.31.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Logging & Monitoring
loguru>=0.7.0

# Development & Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Performance & Optimization
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional ML Libraries (Optional but recommended)
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Security
cryptography>=41.0.0

# File Handling
pillow>=10.0.0
python-magic>=0.4.27